move socket send packet manager from tcp to ..
