+++ aether/transport/low_level/tcp/tcp.h
@@ -34,8 +34,8 @@
 
 #  include "aether/transport/itransport.h"
 #  include "aether/transport/low_level/tcp/data_packet_collector.h"
-#  include "aether/transport/low_level/tcp/socket_packet_send_action.h"
-#  include "aether/transport/low_level/tcp/socket_packet_queue_manager.h"
+#  include "aether/transport/low_level/socket_packet_send_action.h"
+#  include "aether/transport/low_level/socket_packet_queue_manager.h"
 
 #  include "aether/transport/low_level/sockets/tcp_sockets_factory.h"
